Fix all broken specs at once
--- Fixtures were wrong:

a) Users fixtures had no email, and it defaults to empty string (""),
resulting in a NOT UNIQUE constraint violation.
b) Tasks fixtures had a wrong column name (descrption for description)

--- Controller tests

Url helpers were all wrong. Tasks urls are in the projects/ namespace
and so the url helpers start with _project and receive a `Project`
object as argument.

--- Destroying dependent tasks

When destroying a project we weren't destroying the children tasks and
we were encountering a Foreign key constraint violation.

Now the Project `has_many :tasks` association has a `dependent: destroy`
which ensures it runs `task.destroy` for each of the project's tasks
before destroying the project.